Moves tests to proper package and cleans them up
